// src/app/models/ActividadesCronograma.ts
import { ActividadesDiarias } from "./actividades-diarias";
import { Cronograma } from "./Cronograma";

export class ActividadesCronograma {
    id: any;

    numActividad: any;

    fechaSeguimiento: any;

    fechaFinalizacion: any;

    porcentaje: any;

    observacion: any;
    actividadesDiarias:ActividadesDiarias;
    cronograma:Cronograma;

    
}